update dto product

+++ pro/src/main/java/com/pro_servises/pro/controller/OrderController.java
@@ -17,21 +17,21 @@ public class OrderController {
     private OrderServiceImpl orderServiceImpl;
 
     @PostMapping("/add_order")
-    public Order addOrder(@RequestBody Order order,@RequestParam Long product_id) {
+    public Order addOrder(@RequestBody Order order,@RequestParam Integer product_id) {
         return orderServiceImpl.addOrder(order,product_id);
     }
 
     @DeleteMapping("/delete_order/{id}")
-    public void deleteOrder(@PathVariable Long id) {
+    public void deleteOrder(@PathVariable Integer id) {
         orderServiceImpl.deleteOrder(id);
     }
 
     @GetMapping("/gets_order_by_provider_id")
-    public List<Order> getOrderByProviderId(@RequestParam Long providerId) {
+    public List<Order> getOrderByProviderId(@RequestParam Integer providerId) {
         return orderServiceImpl.getAllOrdersByProviderId(providerId);
     }
     @GetMapping("/gets_order_by_user_id")
-    public List<Order> getOrderByUserId(@RequestParam Long userId) {
+    public List<Order> getOrderByUserId(@RequestParam Integer userId) {
         return orderServiceImpl.getAllOrdersByUserId(userId);
     }
